requirements.txt
## install with:
## pip install -r requirements.txt
python-docx
tk
